feat: [SC-24139] Updated AvatarWithTooltip approach moving queries to outside of the comp.
