+++ components/2 - molecules/ens-avatar.tsx
@@ -1,6 +1,8 @@
+/* eslint-disable react-hooks/exhaustive-deps */
 import { Profile } from "@/data/ensProfiles";
 import { AvatarSize, AvatarWithTooltip } from "./avatar-with-tooltip";
 import { getEnsAvatarSrc } from "@/lib/client/avatar";
+import { useEffect, useState } from "react";
 
 interface EnsAvatarProps {
   profile: Profile;
@@ -23,8 +25,36 @@ export const EnsAvatar = ({
   className = "",
   size = AvatarSize.MEDIUM,
 }: EnsAvatarProps) => {
+  const urls = getNameKitAvatarSrcs(profile);
+  const [avatarUrlUsed, setAvatarUrlUsed] = useState(0);
+
+  const [avatarSrc, setAvatarSrc] = useState<string>("");
+
+  const queryAvatarFromNewSource = () => {
+    fetch(urls[avatarUrlUsed])
+      .then(async (response) => {
+        if (response.ok) {
+          setAvatarSrc(urls[avatarUrlUsed]);
+        } else {
+          console.error(response);
+
+          setAvatarUrlUsed(avatarUrlUsed + 1);
+        }
+      })
+      .catch((e) => {
+        console.error(e);
+
+        setAvatarUrlUsed(avatarUrlUsed + 1);
+      });
+  };
+
+  useEffect(() => {
+    queryAvatarFromNewSource();
+  }, [avatarUrlUsed]);
+
   return (
     <AvatarWithTooltip
+      avatarSrc={avatarSrc}
       avatarUrlOptions={getNameKitAvatarSrcs(profile)}
       className={className}
       profile={profile}
